% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_matrix.R
\name{as_matrix}
\alias{as_matrix}
\title{Transforms a tibble into a matrix}
\usage{
as_matrix(tb, long = FALSE, out_type = "numeric")
}
\arguments{
\item{tb}{a tibble}

\item{long}{is the source in long format? Restructures into wide format. Use with caution.}

\item{out_type}{c("numeric", "character"); preferred output data.class. Will introduce NAs from characters if forced to "numeric".}
}
\value{
no return value
}
\description{
This function was excluded from dplyr, but is still usefull in many workflows.
It takes a tibble and transforms it into a matrix. Rownames are taken from the first column, which is subsequently deleted. Colnames are kept.
Columns should be selected before transformation.
}
\examples{
# wide format data
tb_wide <- as_tibble(mtcars,
                     rownames = "model")

tb_wide \%>\%
  as_matrix()
tb_wide \%>\%
  as_matrix(out_type = "character")

# from long format, not recommended
tb_long <- as_tibble(mtcars,
                     rownames = "model") \%>\%
  pivot_longer(2:12)

tb_long \%>\%
  as_matrix(long = TRUE)
}
