% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_intersect_members.R
\name{get_intersect_members}
\alias{get_intersect_members}
\title{Returns membership information on intersections of a VennDiagram}
\usage{
get_intersect_members(x = list())
}
\arguments{
\item{x}{A list of (named) character vectors indicating the different groups.}
}
\value{
a tibble containing BOOLean columns
}
\description{
Best used together with the plotVenn() function.
}
\examples{
# define a (named) list containing three named character vectors
list_4Venn <- list(GroupA = c("Hans", "Jane", "Mark", "Gudrun"),
                   GroupB = c("Jane", "Gudrun", "Simon", "Flo"),
                   GroupC = c("Simon", "Jane", "Theordore", "Jessica"))

plotVenn(list_4Venn)

# get membership information
get_intersect_members(list_4Venn)
}
