% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nin.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Checks which elements of x are not found in y}
\usage{
x \%nin\% y
}
\arguments{
\item{x}{first vector, to be checked}

\item{y}{second vector, checked against}
}
\value{
BOOLean vector of elements of x not in y
}
\description{
Checks which elements of x are not found in y
}
\examples{
x <- c("A", "B", "C", "D")
y <- c("A", "B", "C")

x \%nin\% y
}
