% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_IN.R
\name{\%!in\%}
\alias{\%!in\%}
\title{Checks which elements of x are not found in ydevtool}
\usage{
x \%!in\% y
}
\arguments{
\item{x}{first vector, to be checked}

\item{y}{second vector, checked against}
}
\value{
BOOLean vector of elements of x not in y
}
\description{
Checks which elements of x are not found in ydevtool
}
\examples{
x <- c("A", "B", "C", "D")
y <- c("A", "B", "C")

x \%!in\% y
}
