% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvl.R
\name{lvl}
\alias{lvl}
\title{Returns the factor levels of a column}
\usage{
lvl(tb, column)
}
\arguments{
\item{tb}{a tibble containing the column of interest}

\item{column}{name of the column}
}
\description{
Can be used to get an idea which unique elements a vector contains.
Does not need to be a factor vector.
}
