% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple.R
\name{multiple}
\alias{multiple}
\title{Checks if elements are duplicated}
\usage{
multiple(v)
}
\arguments{
\item{v}{a vector}
}
\value{
a BOOLean vector indicating all elements with duplicates in v
}
\description{
Checks if elements are duplicated
}
\examples{
v <- c("a", "b", "c", "a", "d")
multiple(v)
}
