% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_png.R
\name{open_png}
\alias{open_png}
\title{Opens a ggplot2 plot in an external window}
\usage{
open_png(
  plot,
  width = 1000,
  height = 700,
  legend.title.size = 15,
  legend.text.size = 12
)
}
\arguments{
\item{plot}{A ggplot2 object}
}
\value{
no return value
}
\description{
If R struggles with plotting large and complex plots, this function
exports the plot to a temporary file and opens it in an external
window.
This bypasses the sometimes rather slow plotting by R.
}
\examples{
# plot the displacement of car eniges over the number of cylinders
p <-
  mpg \%>\%
    ggplot(aes(x = cyl,
               y = displ)) +
  geom_bar(stat = "identity") +
  theme_classic()

# export to temporary file and open
open_png(p)
}
