% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_dyn_volcano_cut.R
\name{param_dyn_volcano_cut}
\alias{param_dyn_volcano_cut}
\title{Used to parametrise the dyn_volcano_cut() function}
\usage{
param_dyn_volcano_cut(
  df = data.frame(x = c(1.5, 4), y = c(0.001, 0.05)),
  p_max = 0.1
)
}
\arguments{
\item{df}{a data.frame containing x and y coordinates for the two points the hyperbolic function is forces through}

\item{p_max}{maximal p.value/FDR allowed. Only relevant for very large fold-changes}
}
\value{
list(s, FC_min) s ... "slope" of the hyperbolic function; FC_min ... minimum FC the hyperbolic function converges towards in x direction
}
\description{
Can be used to parametrize the dyn_volcano_cut function if the standard parameters are not sufficient.
The hyperbolic function follows:
}
\details{
y = s / (log2FC - log2FC.min) + FDR.min

s ... "slope" of the function
log2FC ... log2 fold-change of features
log2FC.min ... minimum log2FC towards which the function converges towards in x: lim(f; y > 0) = log2FC.min
FDR.min ... minimum FDR the function converges to in y: lim(f; x > Inf) = FDR.min
}
