% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVenn.R
\name{plotVenn}
\alias{plotVenn}
\title{Plots Venn Diagram}
\usage{
plotVenn(x = list(), shape = "ellipse")
}
\arguments{
\item{x}{A list of (named) character vectors indicating the different groups.}

\item{shape}{Either "ellipse" or "circle". Ellipses are more accurate especially in cases of more than three groups.}
}
\value{
no return value
}
\description{
Plot Venn-diagrams from a list of named character vectors.
This is used to indicate the overlaps in different groups.
Beaware: Accurate solutions are not possible for comparisons of many groups. More than 5 groups is generally not advised.
}
\examples{
# define a (named) list containing three named character vectors
list_4Venn <- list(GroupA = c("Hans", "Jane", "Mark", "Gudrun"),
                   GroupB = c("Jane", "Gudrun", "Simon", "Flo"),
                   GroupC = c("Simon", "Jane", "Theordore", "Jessica"))

plotVenn(list_4Venn)
}
