% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_colour.R
\name{rand_colour}
\alias{rand_colour}
\title{Returns n random colours from all RColorBrewer palettes}
\usage{
rand_colour(n = 20)
}
\arguments{
\item{n}{number of colours to return}
}
\description{
This was adapted from:
https://stackoverflow.com/questions/15282580/how-to-generate-a-number-of-most-distinctive-colors-in-r
User: JelenaCuklina
}
