% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsxglimpse.R
\name{xlsxglimpse}
\alias{xlsxglimpse}
\title{Opens a tibble or data.frame in an Excel Window}
\usage{
xlsxglimpse(df)
}
\arguments{
\item{df|tb, }{data.frame or tibble to be displayed}
}
\value{
no return value
}
\description{
The main purpose of this function is to display data structures in Excel
OR
export data for quick saving.
}
\details{
This function exports a tibble or data.frame to the temporary folder of R
and opens it in an excel-file.
The temporary file folder of R is erased with every clean shut-down, don't forget
to save crucial data.
Also be aware that Excel has problems properly loading and viewing large files.
Use matrix operations or tidy expressions to subset large data sets before exporting to Excel.
}
\examples{
# open the "mpg" data set in an Excel window
library(ggplot2) # contains the mpg data set
print(mpg)

# direct call
xlsxglimpse(mpg)

# tidy implementation
mpg \%>\%
  select(manufacturer, model, cyl, class) \%>\%
  filter(cyl > 4) \%>\%
  xlsxglimpse()
}
